Add initial DETECT screenings
I had to do some data cleaning so that we could add values for all the DETECT screenings that were done between 2019-07-01 and 2020-09-30 to FM Pro. The files I'm committing here save to the number of duplicate MedStar IDs that existed in the data by year and month and also reduces the data to one row per MedStar ID. I then sent the cleaned data to Sunil to import into FM Pro. I will likely need to do a similar process each month.